% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_powerscribe.R
\name{assign_attending_accno}
\alias{assign_attending_accno}
\title{Assigns given accession number to attending}
\usage{
assign_attending_accno(acc_no, attn, fellow_res = "", test = FALSE)
}
\arguments{
\item{acc_no}{Single accession number}

\item{attn}{Attending "Provider Number". Accepts 0 or NULL for unassignment.}

\item{fellow_res}{Fellow or Resident "Provider Number" (not implemented yet)}

\item{test}{Whether to use Powerscribe test or prod API, default is FALSE (prod)}
}
\description{
Returns string "OK" if successful and HTTP response content if not.
}
