% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_powerscribe.R
\name{assign_reports}
\alias{assign_reports}
\title{assign core function}
\usage{
assign_reports(df, ps_test = FALSE)
}
\arguments{
\item{df}{(tibbles) as produced by "prep" functions.}

\item{ps_test}{Whether to use the test or prod PS360 API. Defaults to FALSE}
}
\description{
Essentially just a small wrapper for assign_attending_accno()
Returns a dataframe, last column contains "OK" for successful
assignments and the error message for failures.
}
\examples{
\dontrun{
worklist <- prep_worklist(sim_worklist(1))
ct_att <- sim_shift_rads(1, "CT", "am")
cts <- divide_ct(worklist, ct_att, 2, bypass_ps = TRUE)

ct_att_unnested <- assign_reports(cts[[1]], ps_test = TRUE)
}

}
