% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pins.R
\name{carryover_weights}
\alias{carryover_weights}
\title{Carry over exam weights from prior shift}
\usage{
carryover_weights(
  reader_df,
  test_pin = FALSE,
  service = "Body",
  pin_board = "rsconnect"
)
}
\arguments{
\item{reader_df}{Dataframe from QGenda schedule db}

\item{test_pin}{Logical. Whether to write to production or test pin. Defaults to FALSE = prod.}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}

\item{pin_board}{Defaults to "rsconnect"}
}
\description{
Gets overflow or underassigned numbers from previous
 shift saved in pinned balance sheet and adjusts current
 weights accordingly.
}
