% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_divide_cr.R
\name{divide_cr}
\alias{divide_cr}
\title{Divide radiographs core function}
\usage{
divide_cr(cr_df, atts, bypass_ps)
}
\arguments{
\item{cr_df}{Worklist}

\item{atts}{CT/MRI attendings; dataframe with multiple rows}

\item{bypass_ps}{Bypasses check of PS360 database (kicks out dictated/queued exams). Defaults to FALSE.}
}
\description{
This function assigns up to 7 exams to the primary
CR attending and the rest to attendings on the overflow
list (usually CT and MR) with the highest number of
exams left in the pinned balance sheet.
}
