% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_divide_ct.R
\name{divide_ct}
\alias{divide_ct}
\title{divide_ct core function}
\usage{
divide_ct(
  wl,
  atts,
  bootstrap_samples = 20,
  late_dry = FALSE,
  bypass_ps = FALSE
)
}
\arguments{
\item{wl, atts}{Data frames (tibbles) as produced by "prep" functions.}

\item{bootstrap_samples}{Number of bootstrapped optimization samples. Defaults to 50.}

\item{late_dry}{Don't give wet reads to people on a late shift (parameter "swim" in df.)}

\item{bypass_ps}{Bypasses check of PS360 database (kicks out dictated/queued exams). Defaults to FALSE.}
}
\description{
Creates dataframe with CT's assigned to attendings.
Expects dataframes with specific columns as returned
by the "prep" functions
}
\details{
Returns a list containing two dataframes:
1. Assigned cases
2. Leftovers
and a 2 x n matrix:
3. Bootstrap iteration | Optimization parameter (more = better)

The latter is mostly for academic/testing purposes and can be ignored
in production.
}
\examples{
worklist <- prep_worklist(sim_worklist(1))
ct_att <- sim_shift_rads(1, "CT", "am")
assignments <- divide_ct(worklist, ct_att, 2, FALSE, TRUE)
}
