% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_divide_neuro.R
\name{divide_neuro}
\alias{divide_neuro}
\title{divide_neuro core function}
\usage{
divide_neuro(ctmr, atts, bootstrap_samples = 1, bypass_ps = FALSE)
}
\arguments{
\item{ctmr, atts}{Data frames (tibbles) as produced by "prep" functions.}

\item{bootstrap_samples}{Number of bootstrapped optimization samples. Defaults to 50.}

\item{bypass_ps}{Bypasses check of PS360 database (kicks out dictated/queued exams). Defaults to FALSE.}
}
\description{
Creates dataframe with Neuro CT and MR's assigned to attendings.
Expects dataframes with specific columns as returned
by the "prep" functions.
}
\details{
Returns a list containing two dataframes:
1. Assigned cases
2. Leftovers
and a 2 x n matrix:
3. Bootstrap iteration | Optimization parameter (more = better)

The latter is mostly for academic/testing purposes and can be ignored
in production.

If only 1 bootstrap sample is selected the assignment
will not be done randomly but in the sequence of min_exams.
I.e. attendings with more open points will preferentially be
assigned exams (for cleanup round/crossover reads).
}
\examples{
worklist <- prep_worklist(sim_worklist(1))
mr_att <- sim_shift_rads(1, "MR", "am")
\dontrun{
divide_mr(worklist, mr_att, 5)
}

}
