% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prep_worklists.R
\name{fetch_completed}
\alias{fetch_completed}
\title{Fetch completed exams}
\usage{
fetch_completed(reassign = FALSE, service = "Body")
}
\arguments{
\item{reassign}{Defaults to FALSE = ignores cases previously touched by ignyt.}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}
}
\description{
Gets worklist of completed exams up to current date from RadOps.RadDataView
(includes max. 90 days old to include foreign exams) and filters relevant exams.
Requires database RadOps_UX (username) and RadOps_PW (password) to be set
in GlobalEnv.
}
