% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prep_worklists.R
\name{fetch_scheduled}
\alias{fetch_scheduled}
\title{Fetch scheduled exams}
\usage{
fetch_scheduled()
}
\description{
Gets worklist of scheduled exams today + next 3 days from RadOps.diagnostic_master_v
Requires database RadOps_UX (username) and RadOps_PW (password) to be set
in GlobalEnv.
}
