% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{force_rbind}
\alias{force_rbind}
\title{force_rbind helper function}
\usage{
force_rbind(x, y)
}
\arguments{
\item{x, y}{Data frames; x can be NULL or NA.}
}
\description{
Bind_rows of 2 data frames regardless of NA or NULL in first arg.
Is used to assign cases to nested dataframe when reader
has no cases yet.
}
