% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_smartsheet.R
\name{get_smartsheet}
\alias{get_smartsheet}
\title{Get smartsheet}
\usage{
get_smartsheet(sheet_id, api_key)
}
\arguments{
\item{sheet_id}{ID of spreadsheet}

\item{api_key}{Alphanumeric key to connect to smartsheet API}
}
\value{
Tibble with smartsheet-spreadsheet content
}
\description{
Uses the smartsheet.com API to fetch a given spreadsheet.
Returns a tibble (uses read_csv internally).
}
\keyword{internal}
