% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pins.R
\name{kable_balance_sheet}
\alias{kable_balance_sheet}
\title{Balance Sheet pin -> Kable}
\usage{
kable_balance_sheet(test_pin = FALSE, service = "Body")
}
\arguments{
\item{test_pin}{Logical. Whether to write to production or test pin. Defaults to FALSE = prod.}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}
}
\description{
Gets balance sheet from pin and outputs to nice knitr::kable
table via ignyt_kable()
}
