% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pins.R
\name{null_pin_bal}
\alias{null_pin_bal}
\title{Null outstanding pinned balances}
\usage{
null_pin_bal(test_pin = FALSE, service = "Body")
}
\arguments{
\item{test_pin}{Logical. Whether to write to production or test pin. Defaults to FALSE = prod.}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}
}
\value{
Data frame with new balance sheet.
}
\description{
Nulls positive balances in pin board.
Run by AM report.
}
