% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pins.R
\name{pin_exists}
\alias{pin_exists}
\title{Check if pin name exists}
\usage{
pin_exists(name, board = "rsconnect")
}
\arguments{
\item{name}{Name of pin}

\item{board}{Defaults to "rsconnect" (the default board for ignyt)}
}
\description{
Check if pin name exists
}
