% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_smartsheet.R
\name{pivot_dmts}
\alias{pivot_dmts}
\title{Pivot DMT tables}
\usage{
pivot_dmts(faculty, fellows = NULL, .keep_cols = FALSE)
}
\arguments{
\item{faculty}{Dataframe with faculty members and their CT/MR DMTs}

\item{fellows}{Optional. Table with fellows.}

\item{.keep_cols}{Defaults to FALSE, discards columns not present in internal table.}
}
\description{
Brings DMT tables from wide to long format.
Adds column "mod" with modality code.
Will currently discard all columns that are not
in the built-in tables.
}
\keyword{Internal}
