% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readers.R
\name{prep_readers}
\alias{prep_readers}
\title{Fetch today's QGenda schedule}
\usage{
prep_readers(tomorrow_shift = FALSE, date = NULL, service = "Body")
}
\arguments{
\item{tomorrow_shift}{Whether to get today's or tomorrow's roster. Defaults to today.}

\item{date}{Optional parameter to get readers of a specific date (format YYYY-MM-DD)}

\item{service}{Department service. Defaults to "Body", can be "Neuro" or "Nuclear"}
}
\description{
Connects to the QGenda RadOps SQL table and gets today's readers.
Database username and password default to environment
variables (database-name_UX and _PW).
Returns a dataframe.
}
