% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prep_worklists.R
\name{prep_worklist}
\alias{prep_worklist}
\title{Prepare CT/MR worklist}
\usage{
prep_worklist(worklist, service = "Body")
}
\arguments{
\item{worklist}{Tibble as produced by fetch_*() functions}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}
}
\description{
Adds dmt column to worklist and assigns best/most likely DMT.
Uses internal reference tables (exported in namespace).
General DMT will be assigned to all exams where no other DMT can
reliably be assigned.
Returns tibble of CT and MRI exams with dmt in last column.
}
