% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_powerscribe.R
\name{ps_auth}
\alias{ps_auth}
\title{Powerscribe API authentication}
\usage{
ps_auth(ps_UX, ps_PW, test = FALSE)
}
\arguments{
\item{ps_UX, ps_PW}{Powerscibe API username and password}

\item{test}{If TRUE uses PS test API. Default is FALSE.}
}
\description{
Initializes Powerscribe360 API session. Returns HTTP response code.
Will try to get ps_UX and ps_PW via Sys.getenv if they are left empty.
}
\examples{
\dontrun{
api_response <- ps_auth() # if set previously via Sys.setenv

api_response <- ps_auth(Sys.getenv("myPSusr"), Sys.getenv("myPSpwd"))

# Careful! Clear text password in code is generally discouraged
api_response <- ps_auth("username", "password", test = TRUE)
}

}
