% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_powerscribe.R
\name{ps_get_status}
\alias{ps_get_status}
\title{Powerscribe get status of examination}
\usage{
ps_get_status(acc_nos, test = FALSE)
}
\arguments{
\item{acc_nos}{Accession numbers of examinations.
Can be either a single string separated by commas
or a vector with multiple numbers.}

\item{test}{If TRUE uses PS test API. Default is FALSE.}
}
\description{
Gets status of accession number in Powerscribe.
Originally used as a fallback if the RIS database lags behind
but can be used for other purposes.
Needs an active API session initialized with ps_auth().
Returns a tibble/dataframe.
}
