% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pins.R
\name{register_ignyt_board}
\alias{register_ignyt_board}
\title{Register Ignyt Board}
\usage{
register_ignyt_board(local = FALSE, api_key = Sys.getenv("rsconnect_api_key"))
}
\arguments{
\item{local}{Logical. Whether to use local or remote pinboard. Defaults to remote (FALSE).}

\item{api_key}{RStudio Connect API key. Defaults to environment variable "rsconnect_api_key".}
}
\description{
Reigster ignyt pinboard either locally
or on RSConnect.
Uses environment variable "rsconnect_api_key"
if not otherwise specified
}
