% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readers.R
\name{shift_rads}
\alias{shift_rads}
\title{Prepare shift readers}
\usage{
shift_rads(
  modality = NULL,
  shift,
  date = NULL,
  service = "Body",
  dmt_df = fac_dmts,
  faculty = TRUE,
  pin_test = FALSE,
  only_carryover = FALSE,
  tomorrow = FALSE
)
}
\arguments{
\item{modality}{Can be either "CT" or "MR"}

\item{shift}{Can be either "am", "pm", "Evening Shift" or "Night Shift"}

\item{date}{Optional parameter to get readers of a specific date (format YYYY-MM-DD)}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}

\item{dmt_df}{Reference datarame with dmts for faculty (defaults to fac_dmts provided with package)}

\item{faculty}{Default = TRUE. When FALSE returns CT or MR fellows.}

\item{pin_test}{Whether to use test pin names. Defaults to FALSE.}

\item{only_carryover}{Default = FALSE at the beginning of every shift.
If true does not add any min_exams (used by CR and cleanup)}

\item{tomorrow}{Gets tomorrow's shift (defaults to FALSE = today's shift).
Used for testing in the evening.}
}
\description{
Takes dataframe returned from "prep_readers" and
selects readers for given shift or modality.
Returns a dataframe.
}
