% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testing.R
\name{sim_assign}
\alias{sim_assign}
\title{Simulate assignment}
\usage{
sim_assign(df, n_fail = 0)
}
\arguments{
\item{df}{(tibbles) as produced by "prep" functions.}

\item{n_fail}{Number of cases that should be simulated to fail (picked randomly)}
}
\description{
Assings an "OK" to all exams by default. Number of exams for simulated
failures can be specified as the only parameter.
Expects and returns data frame (analogous to core assign function).
}
