% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testing.R
\name{sim_crs}
\alias{sim_crs}
\title{Get simulated radiographs of a 15 minute window}
\usage{
sim_crs(day_n, time)
}
\arguments{
\item{day_n}{Which day should be selected (1:7)}

\item{time}{What timepoint should be simulated (character with format HH:MM)}
}
\description{
Function returns the worklist on a given day
and time, including examinations the 15 minutes before.
(Assumes that CR Rmd is run in 15 minute intervals and
everything before has been assigned)
}
