% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testing.R
\name{sim_fac_day}
\alias{sim_fac_day}
\alias{sim_fac_day_mod}
\title{Simulate QGenda assignment}
\usage{
sim_fac_day(day_n, mod = NULL)
}
\arguments{
\item{day_n}{Which day should be returned (1:7)}

\item{mod}{Modality to filter by (can be empty)}
}
\description{
Gets QGenda tibble for given day (1:7).
}
