% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_testing.R
\name{sim_worklist}
\alias{sim_worklist}
\title{Get a simulated worklist}
\usage{
sim_worklist(day_n)
}
\arguments{
\item{day_n}{Which day should be returned (1:7), "all" returns whole week.}
}
\description{
Function returns either the worklist of a single day
or of the whole week.
}
