% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slice_min_exams}
\alias{slice_min_exams}
\title{Get min. exams}
\usage{
slice_min_exams(df)
}
\arguments{
\item{df}{data.frame as generated by "divide" core functions}
}
\description{
Returns dataframe reduced to one line per Reader ("Name") with
minimum value in column "min_exams" (requires those two column names).
}
