% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_smartsheet.R
\name{smartsheet_df_check}
\alias{smartsheet_df_check}
\title{Test DMT lists from Smartsheet download}
\usage{
smartsheet_df_check(fac, fell = NULL)
}
\arguments{
\item{fac}{Dataframe with faculty members}

\item{fell}{Dataframe with fellows. Optional.}
}
\description{
Performs various checks to avoid later assignment
errors after updating the DMT list from Smartsheets.
Returns TRUE if all checks are passed successfully.
}
\keyword{Internal}
