% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_transfer_points.R
\name{transfer_points}
\alias{transfer_points}
\title{Transfer assigned exam points between two data.frames}
\usage{
transfer_points(df_a, df_b)
}
\arguments{
\item{df_a}{Dataframe A (origin)}

\item{df_b}{Dataframe B (target)}
}
\description{
Used to transfer the two columns min_exams and sum_assigned
from data.frame A to data.frame B. Column "Name" is used
as a unique identifier.
Use when perfoming multiple assignments of the same reader
within a single document without updating the balance sheet.
}
