% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pins.R
\name{update_pin_bal}
\alias{update_pin_bal}
\title{Update weighted exam balances in pin}
\usage{
update_pin_bal(df, test_pin = FALSE, service = "Body", overwrite = FALSE)
}
\arguments{
\item{df}{Dataframe, needs to have columns "User Name" and "min_exams"}

\item{test_pin}{Logical. Whether to write to production or test pin. Defaults to FALSE = prod.}

\item{service}{Defaults to "Body", can alternatively be "Neuro" or "Nuclear"}

\item{overwrite}{Logical. Whether to overwrite the pin. Defaults to FALSE = add new balance.}
}
\description{
There is a pin used to store overflow and underassignment.
This function takes the unnested dataframe of assigned exams,
boils it down to per-reader and updated the pin.
}
